requirements.txt
requests==2.31.0
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.10.6
python-dotenv==1.0.0
schedule==1.2.0
beautifulsoup4==4.12.3
playwright==1.49.0
python-multipart==0.0.6
aiosmtplib==3.0.1
discord-webhook==1.3.0